+++ Example/ShapeUpExample/NotchedShapes/NotchedRectangleExample.swift
@@ -20,7 +20,7 @@ struct NotchedRectangleExample: View {
                 ])
                 .addingNotches(
                     [
-                        Notch(depth: 50).cornerStyle(.rounded(radius: 10)),
+                        Notch(depth: 50).defaultCornerStyle(.rounded(radius: 10)),
                         nil,
                         Notch(.triangle, position: .relative(0.5), length: .relative(0.2), depth: .relative(0.1)),
                         Notch(depth: 60) {